+++ src/selectors/reportSelectorHelpers.js
@@ -3,7 +3,7 @@ const reportSelectorHelpers = {
     getReportData: reportState => reportState?.results?.data,
     getReportLoading: reportState =>
         reportState?.results?.loading ? reportState.results.loading : false,
-    getReportOptions: reportState => reportState.options
+    getReportOptions: reportState => reportState?.options
 };
 
 export default reportSelectorHelpers;